GammaFunctor
